refactor: convert page to component
